+++ src/son_editor/util/context.py
@@ -2,6 +2,9 @@ from son_editor.app import __main__
 from son_editor.app.database import reset_db
 from son_editor.util.requestutil import CONFIG
 
+# URL
+CATALOGUE_INSTANCE_URL = "http://fg-cn-sandman2.cs.upb.de:4012"
+
 
 # Initializes a test-case context
 def init_test_context():
